requirements.txt
# https://github.com/pyppeteer/pyppeteer
pyppeteer
requests
openpyxl
pytesseract
Pillow
wechatsogou